requirements.txt
pillow
numpy
opencv-python
face_recognition
